requirements.txt
huggingface-hub
openai
python-dotenv
requests
